made a new class
