+++ Converter/src/ConversionRunner.java
@@ -30,7 +30,7 @@ public class ConversionRunner
 					case 2:
 						break;
 					case 3:
-						distanceConversion();
+						
 						break;
 					case 4:
 						timeConversion();
@@ -86,15 +86,5 @@ public class ConversionRunner
 		}
   
   
-  		public static void distanceConversion()
-  			{
-  				System.out.println("What units of distance would you like to convert to?");
-  				System.out.println("1) ft to m");
-  				System.out.println("2) m to ft");
-  				System.out.println("3) mi to km");
-  				System.out.println("4) km to mi");
-  				
-  				Scanner userInput_int = new Scanner (System.in);
-  			
-  			}
+  		
 	}
