+++ Converter/src/DistanceConverter.java
@@ -1,0 +1,17 @@
+import java.util.Scanner;
+
+public class DistanceConverter
+	{
+		public static void distanceConversion()
+  			{
+  				System.out.println("What units of distance would you like to convert to?");
+  				System.out.println("1) ft to m");
+  				System.out.println("2) m to ft");
+  				System.out.println("3) mi to km");
+  				System.out.println("4) km to mi");
+  				
+  				Scanner userInput_int = new Scanner (System.in);
+  				
+  			
+  			}
+	}
